requirements.txt
streamlit
requests
python-dotenv
gtts
streamlit-mic-recorder
langchain-nvidia-ai-endpoints
playwright
